requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
attrs==21.2.0
cerberus==1.3.4
click==8.0.3
eve-swagger==0.1.4
eve==1.1.5
events==0.3
flasgger==0.9.5
flask-cors==3.0.10
flask-login==0.5.0
flask-pymongo==2.3.0
flask==2.0.2
importlib-resources==5.4.0
itsdangerous==2.0.1
jinja2==3.0.2
jsonschema==4.2.1
markupsafe==2.0.1
mistune==0.8.4
pymongo==3.12.1
pyrsistent==0.18.0
python-dotenv==0.19.1
pyyaml==6.0
simplejson==3.17.5
six==1.16.0
werkzeug==2.0.2
zipp==3.6.0
